+++ src/components/show-edit-pod/show-edit-pod.js
@@ -238,7 +238,10 @@ class ShowEditPod extends React.Component {
    * @method content
    */
   get contentProps() {
-    let { className, onEdit, ...props } = this.props;
+    let { ...props } = this.props;
+
+    delete props.onEdit;
+    delete props.className;
 
     if (this.props.onEdit !== false) {
       props.onEdit = this.onEdit;
@@ -253,7 +256,10 @@ class ShowEditPod extends React.Component {
    * @method content
    */
   get editingProps() {
-    let { className, onEdit, ...props } = this.props;
+    let { ...props } = this.props;
+
+    delete props.onEdit;
+    delete props.className;
 
     props.as = 'secondary';